fix(index.js): Resolve an issue where the output file was constructed n times where n is the number
The output file had a buffer instanciated and the full current content generated for each potential
import file found. This had two concequences: 1: The function runs slower than it needed to because
of the extra work an issue that is exacerbated the larger the number of imports. 2: Any following
functions piped after sass-generate-contents would be runmultiple times rather than once slowing
down the rest of the chain. This was resolved by moving the file construction code in to a function
run on stream end.